+++ common/autoinstallers/lint/change-all.ts
@@ -24,7 +24,7 @@ function run() {
       process.exit(1);
     }
 
-    console.log(chalk.green(`[Notice] no message is supplied, we'll use latest commit mesage: ${chalk.red.bold(lastCommitMessage)}`));
+    console.log(chalk.green(`[Notice] no message is supplied, we'll use latest commit message: ${chalk.red.bold(lastCommitMessage)}`));
     message = lastCommitMessage;
   } else {
     const result = spawnSync(