+++ packages/vgrammar-util/src/field.ts
@@ -7,14 +7,35 @@
  * @license
  */
 
+import { isArray, isFunction } from '@visactor/vutils';
 import { accessor } from './accessor';
 import { getter } from './getter';
 import { splitAccessPath } from './splitAccessPath';
-import type { FieldGetterGeneratorOptions } from './types';
+import type { FieldGetterFunction, FieldGetterGeneratorOptions } from './types';
 
-export const field = (fieldStr: string, name?: string, opt: FieldGetterGeneratorOptions = {}) => {
+const fieldSingle = (fieldStr: string | FieldGetterFunction, name?: string, opt: FieldGetterGeneratorOptions = {}) => {
+  if (isFunction(fieldStr)) {
+    return fieldStr;
+  }
+
   const path = splitAccessPath(fieldStr);
   const parsedField = path.length === 1 ? path[0] : fieldStr;
 
   return accessor(((opt && opt.get) || getter)(path), [parsedField], name || parsedField);
+};
+
+export const field = (
+  fieldStr: string | string[] | FieldGetterFunction | FieldGetterFunction[],
+  name?: string,
+  opt: FieldGetterGeneratorOptions = {}
+) => {
+  if (isArray(fieldStr)) {
+    const funcs = fieldStr.map(entry => fieldSingle(entry, name, opt));
+
+    return (datum: any) => {
+      return funcs.map(func => func(datum));
+    };
+  }
+
+  return fieldSingle(fieldStr, name, opt);
 };